pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "butns"
version = "0.1.0"
requires-python = ">=3.13"
[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "black>=24.0",
  "ruff>=0.5.0",
  "python-dotenv>=1.0"
]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
